+++ lib/glimmer/swt/text_proxy.rb
@@ -3,11 +3,14 @@ require 'glimmer/swt/widget_proxy'
 module Glimmer
   module SWT
     class TextProxy < WidgetProxy
-      attr_reader :text, :border
+      attr_reader :text, :border, :left, :center, :right
       
       def initialize(parent, args, block)
         args << :border if args.empty?
         @border = true if args.include?(:border) || args.include?(SWTProxy[:border])
+        @left = true if args.include?(:left) || args.include?(SWTProxy[:left])
+        @center = true if args.include?(:center) || args.include?(SWTProxy[:center])
+        @right = true if args.include?(:right) || args.include?(SWTProxy[:right])
         super(parent, args, block)
       end
 
@@ -49,6 +52,9 @@ module Glimmer
         text_id = id
         text_style = 'min-width: 27px; '
         text_style += 'border: none; ' if !@border
+        text_style += 'text-align: left; ' if @left
+        text_style += 'text-align: center; ' if @center
+        text_style += 'text-align: right; ' if @right
         text_class = name
         options = {type: 'text', id: text_id, style: text_style, class: text_class, value: text_text}
         options = options.merge('disabled': 'disabled') unless @enabled